wsi_tissue_tiler/requirements.txt
numpy==1.24.3
pandas==1.3.4
matplotlib==3.5.2
openslide-python==1.1.2
scikit-image==0.18.1
tqdm==4.62.3
# jupyterlab==3.2.1
Pillow==8.1.2
torch==2.0.1
transformers==4.33.2
#safetensors==0.3.3